Safari-611.1.10/LayoutTests/http/tests/referrer-policy-anchor/origin/same-origin-expected.txt

Tests the behavior of origin referrer policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is expected
PASS successfullyParsed is true

TEST COMPLETE
